# api-service/docs/swagger.yaml
basePath: /
definitions:
  common.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error:
        example: Error message
        type: string
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  request.CreateServerRequest:
    properties:
      description:
        type: string
      host:
        maxLength: 255
        type: string
      hostname:
        maxLength: 255
        type: string
      internal_ip:
        type: string
      ipv6_address:
        type: string
      name:
        maxLength: 64
        type: string
      owner_id:
        type: integer
      resource_group_id:
        type: integer
      ssh_credential_id:
        maxLength: 64
        type: string
      ssh_key:
        type: string
      ssh_password:
        type: string
      ssh_port:
        maximum: 65535
        minimum: 1
        type: integer
      ssh_username:
        description: Direct credential fields for creation (will be stored via credential
          management)
        maxLength: 64
        type: string
    required:
    - host
    - hostname
    - name
    - owner_id
    type: object
...
  request.ServerActionRequest:
    properties:
      action:
        enum:
        - restart
        - shutdown
        - update_agent
        - run_command
        type: string
      params:
        additionalProperties: true
        type: object
      server_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - action
    - server_ids
    type: object
  request.ServerStatusCheckRequest:
    properties:
      check_types:
        items:
          type: string
        type: array
      server_ids:
        items:
          type: integer
        minItems: 1
        type: array
      timeout:
        maximum: 300
        minimum: 5
        type: integer
    required:
    - server_ids
    type: object
...
  request.UpdateServerRequest:
    properties:
      description:
        type: string
      host:
        maxLength: 255
        type: string
      hostname:
        maxLength: 255
        type: string
      internal_ip:
        type: string
      ipv6_address:
        type: string
      name:
        maxLength: 64
        type: string
      resource_group_id:
        type: integer
      ssh_credential_id:
        maxLength: 64
        type: string
      ssh_key:
        type: string
      ssh_password:
        type: string
      ssh_port:
        maximum: 65535
        minimum: 1
        type: integer
      ssh_username:
        description: Direct credential fields for updates (will be stored via credential
          management)
        maxLength: 64
        type: string
    type: object
...
  response.BatchServerStatusResponse:
    properties:
      failed_count:
        type: integer
      results:
        items:
          $ref: '#/definitions/response.ServerStatusCheckResult'
        type: array
      success_count:
        type: integer
    type: object
...
  response.RoleResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      permission_count:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/response.PermissionResponse'
        type: array
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
      user_count:
        type: integer
      users:
        items:
          $ref: '#/definitions/response.UserSimpleResponse'
        type: array
    type: object
...
  response.ServerActionResponse:
    properties:
      action:
        type: string
      message:
        type: string
      server_count:
        type: integer
      status:
        type: string
      task_id:
        type: string
    type: object
  response.ServerAgentResponse:
    properties:
      agent_id:
        type: string
      agent_ip:
        type: string
      agent_port:
        type: integer
      binary_path:
        type: string
      config_path:
        type: string
      container_id:
        type: string
      container_name:
        type: string
      created_at:
        type: string
      deployment_type:
        type: string
      id:
        type: integer
      is_online:
        type: boolean
      last_heartbeat_at:
        type: string
      pull_interval:
        type: integer
      pull_mode:
        type: boolean
      server_id:
        type: integer
      service_name:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
  response.ServerFileDeleteResponse:
    properties:
      deleted_at:
        type: string
      file_path:
        type: string
      message:
        type: string
      server_id:
        type: integer
    type: object
  response.ServerFileUploadResponse:
    properties:
      file_path:
        type: string
      file_size:
        type: integer
      message:
        type: string
      server_id:
        type: integer
      uploaded_at:
        type: string
    type: object
  response.ServerListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/response.ServerResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.ServerResponse:
    properties:
      agent:
        $ref: '#/definitions/response.ServerAgentResponse'
      agent_status:
        type: string
      architecture:
        type: string
      cpu_cores:
        type: integer
      created_at:
        type: string
      description:
        type: string
      disk_total:
        type: integer
      docker_status:
        type: string
      host:
        type: string
      hostname:
        type: string
      id:
        type: integer
      internal_ip:
        type: string
      ipv6_address:
        type: string
      kernel_version:
        type: string
      last_checked_at:
        type: string
      memory_total:
        type: integer
      name:
        type: string
      os_distro:
        type: string
      os_version:
        type: string
      owner:
        allOf:
        - $ref: '#/definitions/response.UserResponse'
        description: Related data
      owner_id:
        type: integer
      resource_group_id:
        type: integer
      ssh_credential_id:
        type: string
      ssh_port:
        type: integer
      ssh_status:
        description: Status information (from Redis cache)
        type: string
      updated_at:
        type: string
    type: object
  response.ServerStatusCheckResult:
    properties:
      checks:
        additionalProperties:
          $ref: '#/definitions/response.ServiceStatusInfo'
        description: ssh/agent/docker status details
        type: object
      error_code:
        type: integer
      error_message:
        type: string
      server_id:
        type: integer
      server_name:
        type: string
      status:
        description: success/failed
        type: string
    type: object
  response.ServiceStatusInfo:
    properties:
      checked_at:
        type: string
      containers_count:
        description: for Docker
        type: integer
      error_code:
        type: integer
      error_message:
        type: string
      images_count:
        description: for Docker
        type: integer
      last_heartbeat:
        description: for Agent
        type: string
      response_time:
        description: for SSH
        type: integer
      status:
        description: connected/online/running etc
        type: string
      uptime:
        description: for Agent
        type: integer
      version:
        description: for Agent/Docker
        type: string
    type: object
...
  response.UserResponse:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      gender:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      language:
        example: zh-CN
        type: string
      last_login_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      last_login_ip:
        example: 192.168.1.1
        type: string
      nickname:
        example: John Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
      roles:
        description: 关联数据
        items:
          $ref: '#/definitions/response.RoleResponse'
        type: array
      signature:
        example: This is my signature
        type: string
      status:
        example: 1
        type: integer
      timezone:
        example: Asia/Shanghai
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      username:
        example: john_doe
        type: string
    type: object
  response.UserSimpleResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
info:
  contact:
    email: support@websoft9.com
    name: Websoft9 Support
    url: https://www.websoft9.com
  description: Enterprise cloud application management platform API service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Websoft9 API Service
  version: "1.0"
paths:
...
  /servers:
    get:
      description: List servers with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Resource group ID filter
        in: query
        name: resource_group_id
        type: integer
      - description: Include deleted servers
        in: query
        name: include_deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ServerListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: List servers
      tags:
      - servers
    post:
      consumes:
      - application/json
      description: Create a new server in the system
      parameters:
      - description: Server creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateServerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ServerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Create a new server
      tags:
      - servers
  /servers/{id}:
    delete:
      description: Delete a server from the system
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Delete server
      tags:
      - servers
    get:
      description: Get server details by ID
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ServerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Get server by ID
      tags:
      - servers
    put:
      consumes:
      - application/json
      description: Update server details
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: Server update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ServerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Update server
      tags:
      - servers
  /servers/{id}/files:
    delete:
      description: Delete a single file from server via SSH
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: File path on server
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ServerFileDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Delete file from server
      tags:
      - servers
    post:
      consumes:
      - multipart/form-data
      description: Upload a single file to server via SSH
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Upload path on server
        in: formData
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ServerFileUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Upload file to server
      tags:
      - servers
  /servers/{id}/files/download:
    get:
      description: Download a single file from server via SSH
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      - description: File path on server
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File content
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Download file from server
      tags:
      - servers
  /servers/{id}/status:
    get:
      description: Check SSH, Agent, Docker status of a single server
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ServerStatusCheckResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Get server status
      tags:
      - servers
  /servers/actions:
    post:
      consumes:
      - application/json
      description: Execute operations on single or multiple servers (restart, shutdown,
        etc.)
      parameters:
      - description: Server action request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ServerActionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ServerActionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Execute server actions
      tags:
      - servers
  /servers/status:
    post:
      consumes:
      - application/json
      description: Check SSH, Agent, Docker status of multiple servers
      parameters:
      - description: Status check request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ServerStatusCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BatchServerStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: Check multiple servers status
      tags:
      - servers
securityDefinitions:
  BearerAuth:
    description: Enter the token with the 'Bearer ' prefix, e.g. 'Bearer abc123'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"